diffAE/src/requirements.txt
torch
diffusers
einops
hydra-core
lightning
fabric
torchvision
matplotlib